+++ Demo/src/main/java/com/braintreepayments/demo/DisplayNonceFragment.java
@@ -35,8 +35,7 @@ public class DisplayNonceFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_display_nonce, container, false);
         nonceString = view.findViewById(R.id.nonce);
         nonceDetails = view.findViewById(R.id.nonce_details);